+++ src/Flare.Abstractions/Css/Classes/Markdown.cs
@@ -7,4 +7,6 @@ public static class Markdown
     public const string Root = "flare-markdown";
     /// <summary>The <c>flare-markdown--highlight</c> CSS class.</summary>
     public const string Highlight = "flare-markdown--highlight";
+    /// <summary>The <c>flare-markdown__table-scroll</c> CSS class.</summary>
+    public const string TableScroll = "flare-markdown__table-scroll";
 }